src/app/frontend/requirements.txt
streamlit
opencv-python-headless
pillow
requests
streamlit_drawable_canvas
uvicorn